sats: provide some 'Serialize' method bodies
